web-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
deepl==1.17.0
ffmpeg-python==0.2.0
